Moved queries in to helper class to assist testing
