+++ HackneyAddressesAPI/Gateways/V1/AddressesGateway.cs
@@ -31,7 +31,7 @@ namespace LBHAddressesAPI.Gateways.V1
             var result = new AddressDetails();
 
             //TODO: Move the query in to a helper so it's in one place!
-            string query = GetAddressesQuery(GlobalConstants.Format.Detailed) + GetSingleAddressClause();
+            string query = QueryBuilder.GetSingleAddress(GlobalConstants.Format.Detailed);
             using (var conn = new SqlConnection(_connectionString))
             {
                 //open connection explicity
@@ -58,8 +58,8 @@ namespace LBHAddressesAPI.Gateways.V1
         {
             var result = new PagedResults<AddressDetails>();            
             var dbArgs = new DynamicParameters();//dynamically add parameters to Dapper query
-            string query =  GetAddressesQuery(GlobalConstants.Format.Detailed) + GetSearchAddressClause(request, true, true, ref dbArgs);
-            string countQuery = GetAddressCountQuery() + GetSearchAddressClause(request, false, false, ref dbArgs);
+            string query = QueryBuilder.GetSearchAddressQuery(GlobalConstants.Format.Detailed, request, true, true, false, ref dbArgs);
+            string countQuery = QueryBuilder.GetSearchAddressQuery(GlobalConstants.Format.Detailed, request, false, false, false, ref dbArgs);
             
             using (var conn = new SqlConnection(_connectionString))
             {
@@ -91,11 +91,13 @@ namespace LBHAddressesAPI.Gateways.V1
         public async Task<PagedResults<AddressDetailsSimple>> SearchSimpleAddressesAsync(SearchAddressRequest request, CancellationToken cancellationToken)
         {
             var result = new PagedResults<AddressDetailsSimple>();
-            //Only add Gazetteer to query if NATIONAL or LOCAL are specified
+
             var dbArgs = new DynamicParameters();
 
-            string query = GetAddressesQuery(GlobalConstants.Format.Simple) + GetSearchAddressClause(request, true, true, ref dbArgs);
-            string countQuery = GetAddressCountQuery() + GetSearchAddressClause(request, false, false, ref dbArgs);
+            string query = QueryBuilder.GetSearchAddressQuery(GlobalConstants.Format.Simple, request, true, true, false, ref dbArgs);
+            string countQuery = QueryBuilder.GetSearchAddressQuery(GlobalConstants.Format.Simple, request, false, false, false, ref dbArgs);
+
+
             using (var conn = new SqlConnection(_connectionString))
             {
                 //open connection explicity
@@ -117,93 +119,6 @@ namespace LBHAddressesAPI.Gateways.V1
         }
 
 
-        private static string GetAddressesQuery(GlobalConstants.Format format)
-        {
-            string query = string.Empty;
-            if (format == GlobalConstants.Format.Detailed)
-            {
-                query = "SELECT LPI_KEY as AddressID,UPRN, USRN, PARENT_UPRN as parentUPRN,LPI_Logical_Status as addressStatus,SAO_TEXT as unitName,UNIT_NUMBER as unitNumber,PAO_TEXT as buildingName,BUILDING_NUMBER as buildingNumber,STREET_DESCRIPTION as street,POSTCODE as postcode,LOCALITY as locality,GAZETTEER as gazetteer,ORGANISATION as commercialOccupier, WARD as ward, POSTTOWN as royalMailPostTown,USAGE_DESCRIPTION as usageClassDescription,USAGE_PRIMARY as usageClassPrimary,BLPU_CLASS as usageClassCode, PROPERTY_SHELL as propertyShell,NEVEREXPORT as isNonLocalAddressInLocalGazetteer,EASTING as easting, NORTHING as northing, LONGITUDE as longitude, LATITUDE as latitude, ";
-                query += " SAO_TEXT as Line1, coalesce(UNIT_NUMBER,'') + ' ' + PAO_TEXT as Line2, BUILDING_NUMBER + ' ' + STREET_DESCRIPTION as Line3, LOCALITY as Line4, POSTTOWN as City ";
-            }
-            else
-            {
-                query = "SELECT SAO_TEXT as Line1, coalesce(UNIT_NUMBER,'') + ' ' + PAO_TEXT as Line2, BUILDING_NUMBER + ' ' + STREET_DESCRIPTION as Line3, LOCALITY as Line4, POSTTOWN as City, Postcode, UPRN, LPI_KEY as AddressID ";
-            }
-            return query;
-        }
-
-        private static string GetAddressCountQuery()
-        {
-            return "SELECT count(1) ";
-        }
-
-        private static string GetSingleAddressClause()
-        {
-            return " FROM dbo.combined_address WHERE LPI_KEY = @key";
-        }
-
         
-        private static string GetSearchAddressClause(SearchAddressRequest request, bool includePaging, bool includeRecompile, ref DynamicParameters dbArgs)
-        {
-            
-            string clause = string.Format(" FROM dbo.combined_address L WHERE BLPU_CLASS NOT LIKE 'P%' ");
-
-            if (!string.IsNullOrEmpty(request.PostCode))
-            {
-                dbArgs.Add("@postcode", request.PostCode.Replace(" ", "") + "%");
-                clause += " AND POSTCODE_NOSPACE LIKE @postcode  ";
-            }
-
-            if (!string.IsNullOrEmpty(request.AddressStatus.ToString())) //AddressStatus/LPI_LOGICAL_STATUS
-            {
-                dbArgs.Add("@addressStatus", GlobalConstants.MapAddressStatus(request.AddressStatus));
-                clause += " AND LPI_LOGICAL_STATUS = @addressStatus ";
-            }
-            if(request.UPRN!=null)
-            {
-                dbArgs.Add("@uprn", request.UPRN);
-                clause += " AND UPRN = @uprn ";
-            }
-
-            if (request.USRN!=null)
-            {
-                dbArgs.Add("@usrn", request.USRN);
-                clause += " AND USRN = @usrn ";
-            }
-
-            if (!string.IsNullOrEmpty(request.PropertyClassPrimary.ToString()))
-            {
-                dbArgs.Add("@primaryClass", request.PropertyClassPrimary.ToString());
-                clause += " AND USAGE_PRIMARY = @primaryClass ";
-            }
-            
-            if(!string.IsNullOrEmpty(request.PropertyClassCode))
-            {
-                dbArgs.Add("@propertyClassCode", request.PropertyClassCode + "%");
-                clause += " AND BLPU_CLASS LIKE @propertyClassCode ";
-            }
-
-            if (request.Gazeteer == GlobalConstants.Gazetteer.Both ? false : true)//Gazetteer
-            {
-                dbArgs.Add("@gazetteer", request.Gazeteer.ToString());
-                clause += " AND Gazetteer = @gazetteer ";
-            }
-
-            if (includePaging)//paging
-            {
-                int page = request.Page;
-                int pageSize = request.PageSize;
-                int lower = 0;
-                lower = page == 0 || page == 1 ? 0 : page * pageSize;
-                // paging so if current page passed in is 1 then we set lower bound to be 0 (0 based index). Otherwise we multiply by the page size
-                clause += " ORDER BY street_description, building_number DESC ";
-                clause += string.Format(" OFFSET {0} ROWS FETCH NEXT {1} ROWS ONLY ", lower, pageSize);
-            }
-            if (includeRecompile)//recompile
-            {
-                clause += " OPTION(RECOMPILE) ";
-            }
-            return clause;
-        }
     }
 }
